DPR-482 extract domain editor and elements into separate files
